Added new encode/decode codes. WE can decode without the need to input LSB and Coordinates.
