--- check_stego.py
@@ -1,8 +1,0 @@
-import cv2
-import numpy as np
-
-orig = cv2.imread("input.bmp")
-encoded = cv2.imread("encoded_output.bmp")
-
-diff = np.sum(orig != encoded)
-print(f"Number of pixels changed: {diff}")
