--- fake_payload_txt.py
@@ -1,2 +1,0 @@
-with open("big_test.txt", "w") as f:
-    f.write("HELLO_WORLD\n" * 1_000_000)  # repeat 1 million times
