+++ hello.txt
@@ -1,0 +1,1 @@
+if you are reading this, you are g**
